requirements.txt
customtkinter~=5.2.2
requests~=2.32.5
pandas~=3.0.0
future~=1.0.0
pip~=26.0.1
attrs~=25.4.0
protobuf~=6.33.5
cryptography~=46.0.4
pillow~=12.1.1
docutils~=0.22.4
setuptools~=80.3.1
packaging~=26.0
pyparsing~=3.3.2
gspread~=6.2.1
gspread-formatting~=1.2.1
lxml~=6.0.2
openpyxl~=3.1.5
html5lib~=1.0